convert to much slicker GCP secrets management than env var jank
