+++ .deployment/prod_environment.yml
@@ -6,6 +6,7 @@ channels:
 dependencies:
   - python ==3.10
   - google-auth
+  - google-cloud-secret-manager
   - rioxarray
   - stackstac
   - gdal
