+++ .deployment/terraform.tfstate.backup
@@ -1,7 +1,7 @@
 {
   "version": 4,
   "terraform_version": "1.6.0",
-  "serial": 110,
+  "serial": 112,
   "lineage": "44075ebe-b91a-1e72-3b17-affdfb63c0f1",
   "outputs": {},
   "resources": [
@@ -161,7 +161,7 @@
                 "region": ""
               }
             ],
-            "tags": null,
+            "tags": {},
             "tags_all": {},
             "unique_id": "AROAYDSGCEDJWK62KL7YN"
           },
@@ -265,6 +265,31 @@
     },
     {
       "mode": "managed",
+      "type": "aws_transfer_ssh_key",
+      "name": "sftp_ssh_key_public",
+      "provider": "provider[\"registry.opentofu.org/hashicorp/aws\"]",
+      "instances": [
+        {
+          "schema_version": 0,
+          "attributes": {
+            "body": "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAACAQCjLx8BlIagwUXjFct86LfLfk1jo2+MHQal7lrw5Yj/xc6ZfgriAXDvV/8BmJApyYtn2x7HMXfdDHml/SGHRYWU7cyJz7ieAweieA9vwJskZ052fnN/N75LArh6Al13vBem68+GaPNsV8scSRVKqT55RJsw/QZV2E7r/7X3Wny2PtQ2t4lwe04hZqFo4e2td/VSMQtoV7pgrpzkeWU1uxXE/ZNXgEbl/HCwlMCETFEIAn5dmWczBGj6Rr5Wt3Xh/dmwzUZiRNnHxUBAzcSMEQBb8wu5Qj68bN9fT8mkY8Vev+6Er2YfbdNjce9OZ+Nanyeia3WOY5WvVpZCaJ6SeDYM4JBn8VMpeziXXpAhRG1D8Gj/e7cWpUphQMblxYgCrgqlAzTG14tHDuFZx/iWMusEhYcluG55bKS8181Xzw1JKmm6eS80qokx7wnQctfj7QVthC/z3ULaEsaZxt7GGV8NOAf4m8YxBbojC5lPsRWFk6gefntE38Z6yAIt/vLNdT/k51wuiJxPpsuT0bS1f8Ag02XlBVe+Q007O01neywUb0QW9jantQuQfFCO6hR8fDAXj284VQziQbPxqlA5nvMAisUMNRSGErnHezgFJ2q1rHHRInBgGok7FqVCF1J6e9c1Mv1+rP6JydD7WiBmJNOOf2hmjydzQVq0xLnbt9mfXw== root@5c051681d26c",
+            "id": "s-ef937729b03041ef8/admin/key-32156f34eee04c969",
+            "server_id": "s-ef937729b03041ef8",
+            "user_name": "admin"
+          },
+          "sensitive_attributes": [],
+          "private": "bnVsbA==",
+          "dependencies": [
+            "aws_iam_role.admin",
+            "aws_transfer_server.tf-sftp-burn-severity",
+            "aws_transfer_user.tf-sftp-burn-severity",
+            "data.aws_iam_policy_document.assume_role"
+          ]
+        }
+      ]
+    },
+    {
+      "mode": "managed",
       "type": "aws_transfer_user",
       "name": "tf-sftp-burn-severity",
       "provider": "provider[\"registry.opentofu.org/hashicorp/aws\"]",
@@ -281,7 +306,7 @@
             "posix_profile": [],
             "role": "arn:aws:iam::557418946771:role/tf-sftp-user-iam-role",
             "server_id": "s-ef937729b03041ef8",
-            "tags": null,
+            "tags": {},
             "tags_all": {},
             "timeouts": null,
             "user_name": "admin"
