+++ .devcontainer/dev.Dockerfile
@@ -16,7 +16,7 @@ COPY . /workspace
 WORKDIR /workspace/.devcontainer
 
 # # Setup SSH keys for tf (temporarily til we sort out secrets / OIDC)
-RUN prebuild/gen_ssh_secrets.sh
+RUN prebuild/setup_secrets.sh
 
 # # Get AWS CLI V2
 RUN prebuild/setup_aws.sh
