+++ app.py
@@ -21,11 +21,12 @@ app.include_router(cog.router, prefix='/cog', tags=["Cloud Optimized GeoTIFF"])
 add_exception_handlers(app, DEFAULT_STATUS_CODES)
 
 # # create an SFTP client instance
-SFTP_HOSTNAME = "s-90987336df8a4faca.server.transfer.us-east-2.amazonaws.com"
+SFTP_HOSTNAME = "s-ef937729b03041ef8.server.transfer.us-east-2.amazonaws.com"
 SFTP_USERNAME = "sftp-admin"
 S3_BUCKET_NAME = "burn-severity"
 # sftp_client = SFTPClient(SFTP_HOSTNAME, SFTP_USERNAME, get_ssh_secret())
-sftp_client = SFTPClient(SFTP_HOSTNAME, SFTP_USERNAME, os.getenv('SFTP_SSH_KEY_PRIVATE'))
+SSH_SECRET = os.getenv('SFTP_SSH_KEY_PRIVATE')
+sftp_client = SFTPClient(SFTP_HOSTNAME, SFTP_USERNAME, SSH_SECRET)
 
 @app.get("/")
 def index():
